Update UDPclient: create a function (at client side only)that will attempt to send and receive a message a number of times until success
The arguments to this function include the socket to be used in sending and receiveing a packet, the IP address and port of the destiantion (the server), and the message/packet to be sent. The function returns the response message/packet received. The function uses a loop in which it configures an initial timeout, sends the packet, and waits for a response (blocks on receive).
If there is a timeout, which is an exception, then
- a counter will be increased
- the timeout value will be doubled, restart timer
- print an error message
- retransmit the message
- wait for a response